// src/app/interfaces/apiInterfaces/creditsInterfaces.ts
export interface MovieCreditsResponse {
  id: number;
  cast: MovieCast[];
  crew: MovieCrew[];
}

export interface MovieCast {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: string | null;
  cast_id: number;
  character: string;
  credit_id: string;
  order: number;
}

export interface MovieCrew {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: string | null;
  credit_id: string;
  department: string;
  job: string;
}

// TV --------------------------------------------------------------

export interface TvCreditsResponse {
  id: number;
  cast: TvCast[];
  crew: TvCrew[];
}

export interface TvCast {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: string | null;
  character: string;
  credit_id: string;
  order: number;
}

export interface TvCrew {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: string | null;
  credit_id: string;
  department: string;
  job: string;
}

export interface CreditsQueryParams {
  showId: number;
  showType: "movie" | "tv";
  lang?: string;
}