// src/app/interfaces/apiInterfaces/searchMovieInterfaces.ts
export interface SearchMovie {
  adult: boolean;
  backdrop_path: string;
  genre_ids: number[];
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}

export interface SearchMovieResponse {
  page: number;
  results: SearchMovie[];
  total_pages: number;
  total_results: number;
}

//////////
export interface SearchMovieQueryParams {
  query: string;
  page: number;
  lang?: string;
}